producer/requirements.txt
newsapi-python >= 0.2.7
kafka-python >= 2.0.2
